solve another logout bug
