+++ pages/account/register-trees.tsx
@@ -1,11 +1,10 @@
 import React, { ReactElement } from 'react';
 import RegisterTrees from '../../src/features/user/UserProfile/components/RegisterTrees';
 import { useAuth0 } from '@auth0/auth0-react';
-import AccountHeader from '../../src/features/common/Layout/Header/accountHeader';
+import AccountHeader from '../../src/features/common/Layout/Header/AccountHeader';
 import i18next from '../../i18n';
 import { UserPropsContext } from '../../src/features/common/Layout/UserPropsContext';
 import AccountFooter from '../../src/features/common/Layout/Footer/accountFooter';
-import NewAccountHeader from '../../src/features/common/Layout/Header/newAccountHeader';
 interface Props {}
 const { useTranslation } = i18next;
 export default function Register({}: Props): ReactElement {
@@ -35,7 +34,7 @@ export default function Register({}: Props): ReactElement {
   const { t } = useTranslation(['me']);
   return (
     <>
-      <NewAccountHeader page={'register-trees'} title={t('me:registerTrees')} />
+      <AccountHeader page={'register-trees'} title={t('me:registerTrees')} />
       <div style={{ display: 'flex', justifyContent: 'center' }}>
         {!isLoading && currentUserSlug ? (
           <RegisterTrees
