+++ src/features/common/Layout/Header/AccountHeader.module.scss
@@ -1,82 +1,5 @@
 @import './../../../../theme/theme';
 
-.headerBG {
-  height: 170px;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  position: absolute;
-  top: 0;
-  width: 100%;
-  z-index: 100;
-  background: linear-gradient(
-      to top,
-      rgba(0, 0, 0, 0.5),
-      rgba(0, 0, 0, 0.2),
-      rgba(0, 0, 0, 0),
-      rgba(0, 0, 0, 0)
-    ),
-    url('https://cdn.plant-for-the-planet.org/media/images/app/bg_layer.jpg')
-      center center,
-    $primaryColor;
-  .accountsHeader {
-    height: 100%;
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    width: 100%;
-    max-width: 1200px;
-    padding-top: 10px;
-    position: relative;
-  }
-}
-.accountsTitleContainer {
-  margin-top: 50px;
-  display: flex;
-  justify-content: flex-start;
-  flex-direction: column;
-  .accountsTitle {
-    margin-left: 20px;
-    color: #fff;
-    font-size: 36px;
-    font-weight: bold;
-  }
-}
-.navbar {
-  display: flex;
-  flex-direction: row;
-  justify-content: flex-start;
-  align-items: center;
-  padding-bottom: 12px;
-  overflow-x: scroll;
-  -ms-overflow-style: none; /* IE and Edge */
-  scrollbar-width: none; /* Firefox */
-  padding-right: 10px;
-  > .menuItem {
-    margin: 10px 15px;
-    color: #fff;
-    font-weight: 700;
-    cursor: pointer;
-    min-width: fit-content;
-    > .active {
-      background-color: #fff;
-      border-top-left-radius: 2px;
-      border-top-right-radius: 2px;
-      height: 3px;
-      width: 100%;
-      margin-top: 5px;
-    }
-    > .inActive {
-      height: 3px;
-      margin-top: 5px;
-    }
-  }
-}
-.navbar::-webkit-scrollbar {
-  display: none;
-}
-
 .navContainer {
   padding: 0px 20px;
   display: flex;
@@ -84,8 +7,8 @@
   justify-content: space-between;
   align-items: center;
   position: absolute;
-  top:100px;
-  left:0;
+  top: 100px;
+  left: 0;
   width: 100%;
 }
 .backButton {
@@ -104,16 +27,6 @@
   &:hover {
     cursor: pointer;
   }
-}
-
-@include mdTabletView {
-  .headerBG {
-    height: 320px;
-    .accountsHeader {
-      padding-top: 100px;
-    }
-  }
-  
 }
 
 // new account header
@@ -138,66 +51,66 @@
   z-index: 20;
   width: 100%;
   transition: height 0.5s;
-    > .navigationContainer {
-      padding: 15px;
-      width: 100%;
-      max-width: 1200px;
-      display: flex;
-      flex-direction: row;
-      justify-content: space-between;
-      align-items: center;
-    }
+  > .navigationContainer {
+    padding: 15px;
+    width: 100%;
+    max-width: 1200px;
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    align-items: center;
+  }
 
-    > .titleContainer {
-      padding: 15px;
-      width: 100%;
-      max-width: 1200px;
-      display: flex;
-      flex-direction: row;
-      justify-content: flex-start;
-      align-items: center;
-      > .title {
-        color: #fff;
-        font-size: 36px;
-        font-weight: bold;
+  > .titleContainer {
+    padding: 15px;
+    width: 100%;
+    max-width: 1200px;
+    display: flex;
+    flex-direction: row;
+    justify-content: flex-start;
+    align-items: center;
+    > .title {
+      color: #fff;
+      font-size: 36px;
+      font-weight: bold;
+    }
+  }
+
+  > .menuContainer {
+    width: 100%;
+    max-width: 1200px;
+    display: flex;
+    flex-direction: row;
+    justify-content: flex-start;
+    align-items: center;
+    padding-bottom: 12px;
+    overflow-x: scroll;
+    -ms-overflow-style: none; /* IE and Edge */
+    scrollbar-width: none; /* Firefox */
+    padding-right: 10px;
+    > .menuItem {
+      margin: 10px 15px;
+      color: #fff;
+      font-weight: 700;
+      cursor: pointer;
+      min-width: fit-content;
+      > .active {
+        background-color: #fff;
+        border-top-left-radius: 2px;
+        border-top-right-radius: 2px;
+        height: 3px;
+        width: 100%;
+        margin-top: 5px;
+      }
+      > .inActive {
+        height: 3px;
+        margin-top: 5px;
       }
     }
-    
-    > .menuContainer {
-      width: 100%;
-      max-width: 1200px;
-      display: flex;
-      flex-direction: row;
-      justify-content: flex-start;
-      align-items: center;
-      padding-bottom: 12px;
-      overflow-x: scroll;
-      -ms-overflow-style: none; /* IE and Edge */
-      scrollbar-width: none; /* Firefox */
-      padding-right: 10px;
-      > .menuItem {
-        margin: 10px 15px;
-        color: #fff;
-        font-weight: 700;
-        cursor: pointer;
-        min-width: fit-content;
-        > .active {
-          background-color: #fff;
-          border-top-left-radius: 2px;
-          border-top-right-radius: 2px;
-          height: 3px;
-          width: 100%;
-          margin-top: 5px;
-        }
-        > .inActive {
-          height: 3px;
-          margin-top: 5px;
-        }
-      }
-    }
-    >.menuContainer::-webkit-scrollbar {
-      display: none;
-    }
+  }
+  > .menuContainer::-webkit-scrollbar {
+    display: none;
+  }
 }
 
 @include smTabletView {
