--- src/features/user/Account/styles/AccountNavbar.module.scss
@@ -1,451 +1,0 @@
-@import './../../../../theme/theme';
-
-.accountsPage {
-  width: 100%;
-  display: flex;
-  justify-content: center;
-  flex-direction: column;
-  position: relative;
-  background-color: #fff;
-}
-
-.topLoader {
-  position: absolute;
-  top: 0;
-  width: 100%;
-  height: 5px;
-
-  z-index: 1031;
-}
-
-.topLoaderProgress {
-  width: 0%;
-  background-color: $primaryColor;
-}
-
-.accountsPageContainer {
-  align-self: center;
-  width: 100%;
-  max-width: 1200px;
-  position: relative;
-}
-
-.accountsTitleContainer {
-  display: flex;
-  justify-content: flex-start;
-  width: 980px;
-  .accountsTitle {
-    margin-left: 20px;
-    color: #fff;
-    font-size: 36px;
-    font-weight: bold;
-    padding-bottom: 40px;
-  }
-}
-
-.contentContainer {
-  display: flex;
-  flex-direction: row;
-  gap: 20px;
-}
-
-// accounts
-.accountsContainer {
-  flex-grow: 1;
-  display: flex;
-  flex-direction: column;
-  background: transparent;
-}
-
-.notFound {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  width: 100%;
-  height: 160px;
-  > p {
-    margin-top: 10px;
-  }
-}
-
-// Single record which is not expanded
-
-.singleRecord {
-  padding: 20px;
-  background-color: $light;
-  position: relative;
-}
-
-.borderBottom {
-  border-bottom: #dddbda 1px solid;
-  position: absolute;
-  bottom: 0;
-  width: calc(100% - 40px);
-  left: auto;
-  right: auto;
-}
-
-.recordHeader {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  &:hover {
-    cursor: pointer;
-  }
-  padding: 20px 0;
-  border-bottom: #dddbda 1px solid;
-}
-
-.recordHeaderDesktop {
-  display: none;
-  flex-direction: row;
-  justify-content: space-between;
-  &:hover {
-    cursor: pointer;
-  }
-  padding-bottom: 20px;
-  border-bottom: #dddbda 1px solid;
-}
-
-.selectedRecord {
-  border-radius: 10px;
-  box-shadow: 1px 2px 8px -6px #707070;
-
-  > .borderBottom {
-    display: none;
-  }
-}
-
-.closeRecord {
-  display: block;
-}
-
-.recordDownloads {
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-  align-items: flex-start;
-  gap: 20px;
-  margin-top: 30px;
-  .detail {
-    a {
-      color: $primaryColor;
-      text-decoration: none;
-      font-weight: bold;
-    }
-  }
-}
-
-.recordDataContainer {
-  height: 0px;
-  display: none;
-}
-.recordDataContainer.recordDataContainerSelected {
-  display: none;
-  height: fit-content;
-}
-
-.recordDetails {
-  padding-top: 20px;
-  margin-top: 20px;
-  grid-row-gap: 20px;
-  grid-template-columns: 32% 32% 32%;
-  .detailContainer {
-    display: flex;
-    flex-direction: column;
-    gap: 10px;
-    .detailTitle {
-      font-weight: bold;
-    }
-    a {
-      color: $primaryColor;
-      text-decoration: none;
-      font-weight: bold;
-    }
-  }
-  transition: all 1s ease;
-  -webkit-transition: all 1s ease;
-  -moz-transition: all 1s ease;
-  -ms-transition: all 1s ease;
-  -o-transition: all 1s ease;
-}
-
-.treesDate {
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-}
-.treesDonated {
-  font-weight: bold;
-}
-.recordStatus {
-  font-weight: bold;
-}
-
-.statusAmount {
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-}
-
-// Filters for the records
-
-// .filterContainer {
-//   background-color: $light;
-//   border-radius: 10px;
-//   padding: 20px;
-//   max-width: 300px;
-//   box-shadow: 1px 2px 8px -6px #707070;
-//   height: fit-content;
-//   position: absolute;
-//   right: 0;
-//   top: 120px;
-// }
-
-.filterContainer {
-  background-color: $light;
-  display: flex;
-  flex-direction: row;
-  justify-content: flex-start;
-  align-items: center;
-  overflow-x: scroll;
-  padding: 20px;
-  height: fit-content;
-  -ms-overflow-style: none; /* IE and Edge */
-  scrollbar-width: none; /* Firefox */
-}
-
-.filterContainer::-webkit-scrollbar {
-  display: none;
-}
-
-.filterContainerDesktop {
-  display: none;
-}
-
-.filterTitle {
-  font-weight: 600;
-}
-
-.filterButtons {
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-}
-
-.multiSelectInput {
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
-  background-color: $light;
-  transition-duration: 0.4;
-  transition-timing-function: ease-in;
-  margin-right: 20px;
-  padding: 8px 0;
-  border-radius: 100px;
-  border: 1px solid $primaryColor;
-  min-width: 100px;
-  white-space: nowrap;
-  font-family: $primaryFontFamily;
-}
-
-.multiSelectInputCheckTrue {
-  border: none;
-  padding: 9px 13px;
-  background-color: rgba(104, 176, 48, 0.12);
-}
-
-// .multiSelectInput:hover {
-//   cursor: pointer;
-//   transform: scale(1.05);
-// }
-
-.filterGrid {
-  margin-top: 20px;
-  width: 220px;
-  display: grid;
-  grid-template-columns: repeat(2, 1fr);
-  // grid-template-rows: 1fr;
-  grid-column-gap: 20px;
-  grid-row-gap: 20px;
-}
-
-.filterButtons > button {
-  border-bottom: #dddbda 1px solid;
-  margin-top: 8px;
-  cursor: pointer;
-}
-
-// Single record expanded
-
-.historyContainer {
-  display: flex;
-  justify-content: center;
-  width: 100%;
-}
-
-.amtStatus {
-  display: flex;
-  flex-direction: column;
-  margin-top: 20px;
-}
-.moneyCurr {
-  display: flex;
-  flex-direction: row;
-  margin-bottom: 10px;
-  font-weight: bold;
-}
-
-.project {
-  display: flex;
-  flex-direction: column;
-  margin-top: 30px;
-  width: 150px;
-}
-.projectText {
-  font-weight: bold;
-  margin-bottom: 14px;
-}
-.projectName {
-  color: #68b030;
-}
-.secondRow {
-  display: flex;
-  flex-direction: row;
-  margin-bottom: 24px;
-}
-
-.closeButton {
-  padding: 10px;
-  > svg {
-    width: 20px;
-  }
-}
-
-.optionsRow {
-  display: flex;
-  flex-direction: row;
-  justify-content: flex-end;
-  align-items: center;
-}
-
-.modal {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  overflow-y: auto;
-  width: 100%;
-  height: 100%;
-  background-color: #fff;
-  -ms-overflow-style: none;
-  scrollbar-width: none;
-}
-
-.expandedRecord {
-  height: 100%;
-  width: 100%;
-  padding: 20px;
-}
-
-.recordContainer {
-  padding: 20px 0;
-  display: grid;
-  grid-template-columns: 1fr;
-  grid-column-gap: 20px;
-  grid-row-gap: 20px;
-  .detailContainer {
-    display: flex;
-    flex-direction: column;
-    gap: 10px;
-    .detailTitle {
-      font-weight: bold;
-    }
-    a {
-      color: $primaryColor;
-      text-decoration: none;
-      font-weight: bold;
-    }
-  }
-}
-
-@include smTabletView {
-  .headerBG {
-    height: 300px;
-  }
-  .accountsHeader {
-    margin-top: 80px;
-  }
-  .recordContainer {
-    grid-template-columns: repeat(2, 1fr);
-  }
-}
-
-@include mdTabletView {
-  .contentContainer {
-    margin-top: 30px;
-  }
-
-  .selectedRecord {
-    margin-bottom: 20px;
-    background-color: #fafafa;
-  }
-
-  .recordDataContainer.recordDataContainerSelected {
-    display: block;
-  }
-
-  .recordDetails {
-    display: grid;
-  }
-
-  .recordDownloads {
-    flex-direction: row;
-  }
-
-  .modal {
-    display: none;
-  }
-
-  .recordHeaderDesktop {
-    display: flex;
-  }
-
-  .recordHeader {
-    display: none;
-  }
-}
-
-@include lgLaptopView {
-  .filterContainer {
-    display: none;
-  }
-
-  .closeRecord {
-    display: none;
-  }
-
-  .filterContainerDesktop {
-    display: block;
-    position: relative;
-    background-color: $light;
-    border-radius: 10px;
-    padding: 20px;
-    max-width: 300px;
-    box-shadow: 1px 2px 8px -6px #707070;
-    height: fit-content;
-    right: 0;
-  }
-  .optionsRow {
-    display: none;
-  }
-  .multiSelectInput {
-    // margin-top: 10px;
-    margin-right: 0px;
-  }
-}
-
-:export {
-  primaryColor: $primaryColor;
-}
